August 15th 2025 release.

+++ L2J_Mobius_01.0_Ertheia/java/org/l2jmobius/gameserver/model/actor/Attackable.java
@@ -86,31 +86,38 @@ public class Attackable extends Npc
 	// Raid
 	private boolean _isRaid = false;
 	private boolean _isRaidMinion = false;
+	
 	//
 	private boolean _champion = false;
 	private final Map<Creature, AggroInfo> _aggroList = new ConcurrentHashMap<>();
 	private boolean _canReturnToSpawnPoint = true;
 	private boolean _seeThroughSilentMove = false;
+	
 	// Manor
 	private boolean _seeded = false;
 	private Seed _seed = null;
 	private int _seederObjId = 0;
 	private final AtomicReference<ItemHolder> _harvestItem = new AtomicReference<>();
+	
 	// Spoil
 	private int _spoilerObjectId;
 	private boolean _plundered = false;
 	private final AtomicReference<Collection<ItemHolder>> _sweepItems = new AtomicReference<>();
+	
 	// Over-hit
 	private boolean _overhit;
 	private double _overhitDamage;
 	private Creature _overhitAttacker;
+	
 	// Command channel
 	private CommandChannel _firstCommandChannelAttacked = null;
 	private CommandChannelTimer _commandChannelTimer = null;
 	private long _commandChannelLastAttack = 0;
+	
 	// Soul crystal
 	private boolean _absorbed;
 	private final Map<Integer, AbsorberInfo> _absorbersList = new ConcurrentHashMap<>();
+	
 	// Misc
 	private boolean _mustGiveExpSp;
 	
@@ -256,6 +263,7 @@ public class Attackable extends Npc
 				master.getMinionList().onAssist(this, attacker);
 			}
 		}
+		
 		// Reduce the current HP of the Attackable and launch the doDie Task if necessary
 		super.reduceCurrentHp(value, attacker, skill, isDOT, directlyToHp, critical, reflect);
 	}
@@ -438,6 +446,7 @@ public class Attackable extends Npc
 						totalMemberDamage += memberAggro.getDamage();
 					}
 				}
+				
 				container.damage = totalMemberDamage;
 				
 				if (!partyContainerStream.isPresent())
@@ -612,8 +621,10 @@ public class Attackable extends Npc
 										{
 											finalExp *= attacker.getStat().getExpBonusMultiplier();
 										}
+										
 										clan.addHuntingPoints(attacker, this, finalExp);
 									}
+									
 									if (useVitalityRate())
 									{
 										attacker.updateVitalityPoints(getVitalityPoints(attacker.getLevel(), exp, _isRaid), true, false);
@@ -632,6 +643,7 @@ public class Attackable extends Npc
 						
 						// Get all Creature that can be rewarded in the party
 						final List<Player> rewardedMembers = new ArrayList<>();
+						
 						// Go through all Player in the party
 						final List<Player> groupMembers = attackerParty.isInCommandChannel() ? attackerParty.getCommandChannel().getMembers() : attackerParty.getMembers();
 						for (Player partyPlayer : groupMembers)
@@ -664,6 +676,7 @@ public class Attackable extends Npc
 										}
 									}
 								}
+								
 								rewards.remove(partyPlayer); // Remove the Player from the Attackable rewards
 							}
 							else if (calculateDistance3D(partyPlayer) < Config.ALT_PARTY_RANGE)
@@ -798,7 +811,7 @@ public class Attackable extends Npc
 				getAI().notifyAction(Action.ATTACKED, attacker);
 				
 				// Calculate the amount of hate this attackable receives from this attack.
-				double hateValue = (damage * 100) / (getLevel() + 7);
+				long hateValue = ((long) damage * 100) / (getLevel() + 7);
 				if (skill == null)
 				{
 					hateValue *= attacker.getStat().getMul(Stat.HATE_ATTACK, 1);
@@ -1026,6 +1039,7 @@ public class Attackable extends Npc
 				break;
 			}
 		}
+		
 		if (found)
 		{
 			result.add(secondMostHated);
@@ -1052,6 +1066,7 @@ public class Attackable extends Npc
 			
 			result.add(ai.getAttacker());
 		}
+		
 		return result;
 	}
 	
@@ -1094,6 +1109,7 @@ public class Attackable extends Npc
 			ai.stopHate();
 			return 0;
 		}
+		
 		return ai.getHate();
 	}
 	
@@ -1140,6 +1156,7 @@ public class Attackable extends Npc
 					for (ItemHolder drop : deathItems)
 					{
 						final ItemTemplate item = ItemData.getInstance().getTemplate(drop.getId());
+						
 						// Check if the autoLoot mode is active
 						if (Config.AUTO_LOOT_ITEM_IDS.contains(item.getId()) || isFlying() || (!item.hasExImmediateEffect() && ((!_isRaid && Config.AUTO_LOOT) || (_isRaid && Config.AUTO_LOOT_RAIDS))))
 						{
@@ -1151,6 +1168,7 @@ public class Attackable extends Npc
 							{
 								SkillCaster.triggerCast(mainDamageDealer, null, skillHolder.getSkill(), null, false);
 							}
+							
 							mainDamageDealer.broadcastInfo(); // ? check if this is necessary
 						}
 						else
@@ -1162,6 +1180,7 @@ public class Attackable extends Npc
 							}
 						}
 					}
+					
 					deathItems.clear();
 				}
 			}
@@ -1180,6 +1199,7 @@ public class Attackable extends Npc
 			for (ItemHolder drop : deathItems)
 			{
 				final ItemTemplate item = ItemData.getInstance().getTemplate(drop.getId());
+				
 				// Check if the autoLoot mode is active
 				if (Config.AUTO_LOOT_ITEM_IDS.contains(item.getId()) || isFlying() || (!item.hasExImmediateEffect() && ((!_isRaid && Config.AUTO_LOOT) || (_isRaid && Config.AUTO_LOOT_RAIDS))) || (item.hasExImmediateEffect() && Config.AUTO_LOOT_HERBS))
 				{
@@ -1200,6 +1220,7 @@ public class Attackable extends Npc
 					broadcastPacket(sm);
 				}
 			}
+			
 			deathItems.clear();
 		}
 	}
@@ -1258,6 +1279,7 @@ public class Attackable extends Npc
 				lootItems.add(ItemData.getInstance().getTemplate(item.getId()));
 			}
 		}
+		
 		return lootItems;
 	}
 	
@@ -1292,8 +1314,10 @@ public class Attackable extends Npc
 			{
 				attacker.sendPacket(SystemMessageId.THE_CORPSE_IS_TOO_OLD_THE_SKILL_CANNOT_BE_USED);
 			}
+			
 			return true;
 		}
+		
 		return false;
 	}
 	
@@ -1310,8 +1334,10 @@ public class Attackable extends Npc
 			{
 				sweeper.sendPacket(SystemMessageId.THERE_ARE_NO_PRIORITY_RIGHTS_ON_A_SWEEPER);
 			}
+			
 			return false;
 		}
+		
 		return true;
 	}
 	
@@ -1343,6 +1369,7 @@ public class Attackable extends Npc
 			_overhitAttacker = null;
 			return;
 		}
+		
 		overhitEnabled(true);
 		_overhitDamage = overhitDmg;
 		_overhitAttacker = attacker;
@@ -1491,10 +1518,12 @@ public class Attackable extends Npc
 						break;
 					}
 				}
+				
 				xp *= mul;
 				sp *= mul;
 			}
 		}
+		
 		return new double[]
 		{
 			xp,
@@ -1588,6 +1617,7 @@ public class Attackable extends Npc
 			{
 				_champion = true;
 			}
+			
 			if (Config.SHOW_CHAMPION_AURA)
 			{
 				setTeam(_champion ? Team.RED : Team.NONE, false);
@@ -1699,6 +1729,7 @@ public class Attackable extends Npc
 			{
 				count += diff;
 			}
+			
 			_harvestItem.set(new ItemHolder(_seed.getCropId(), count * Config.RATE_DROP_MANOR));
 		}
 	}
@@ -1898,20 +1929,25 @@ public class Attackable extends Npc
 			{
 				_aggroList.remove(target);
 			}
+			
 			if (_aggroList.isEmpty())
 			{
 				if (getAI() instanceof AttackableAI)
 				{
 					((AttackableAI) getAI()).setGlobalAggro(-25);
 				}
+				
 				if (!isFakePlayer())
 				{
 					setWalking();
 				}
+				
 				clearAggroList();
 			}
+			
 			getAI().setIntention(Intention.ACTIVE);
 		}
+		
 		super.setTarget(object);
 	}
 }
