+++ L2J_Mobius_07.0_PreludeOfWar/java/org/l2jmobius/gameserver/model/actor/Attackable.java
@@ -84,28 +84,34 @@ public class Attackable extends Npc
 	// Raid
 	private boolean _isRaid = false;
 	private boolean _isRaidMinion = false;
+	
 	//
 	private boolean _champion = false;
 	private final Map<Creature, AggroInfo> _aggroList = new ConcurrentHashMap<>();
 	private boolean _canReturnToSpawnPoint = true;
 	private boolean _seeThroughSilentMove = false;
+	
 	// Manor
 	private boolean _seeded = false;
 	private Seed _seed = null;
 	private int _seederObjId = 0;
 	private final AtomicReference<ItemHolder> _harvestItem = new AtomicReference<>();
+	
 	// Spoil
 	private int _spoilerObjectId;
 	private boolean _plundered = false;
 	private final AtomicReference<Collection<ItemHolder>> _sweepItems = new AtomicReference<>();
+	
 	// Over-hit
 	private boolean _overhit;
 	private double _overhitDamage;
 	private Creature _overhitAttacker;
+	
 	// Command channel
 	private CommandChannel _firstCommandChannelAttacked = null;
 	private CommandChannelTimer _commandChannelTimer = null;
 	private long _commandChannelLastAttack = 0;
+	
 	// Misc
 	private boolean _mustGiveExpSp;
 	
@@ -251,6 +257,7 @@ public class Attackable extends Npc
 				master.getMinionList().onAssist(this, attacker);
 			}
 		}
+		
 		// Reduce the current HP of the Attackable and launch the doDie Task if necessary
 		super.reduceCurrentHp(value, attacker, skill, isDOT, directlyToHp, critical, reflect);
 	}
@@ -433,6 +440,7 @@ public class Attackable extends Npc
 						totalMemberDamage += memberAggro.getDamage();
 					}
 				}
+				
 				container.damage = totalMemberDamage;
 				
 				if (!partyContainerStream.isPresent())
@@ -606,6 +614,7 @@ public class Attackable extends Npc
 										attacker.updateVitalityPoints(getVitalityPoints(attacker.getLevel(), exp, _isRaid), true, false);
 										PcCafePointsManager.getInstance().givePcCafePoint(attacker, exp);
 									}
+									
 									if (attacker.getSymbolSealPoints() > 0)
 									{
 										attacker.setSymbolSealPoints(attacker.getSymbolSealPoints() - Config.CONSUME_SYMBOL_SEAL_POINTS);
@@ -627,6 +636,7 @@ public class Attackable extends Npc
 						
 						// Get all Creature that can be rewarded in the party
 						final List<Player> rewardedMembers = new ArrayList<>();
+						
 						// Go through all Player in the party
 						final List<Player> groupMembers = attackerParty.isInCommandChannel() ? attackerParty.getCommandChannel().getMembers() : attackerParty.getMembers();
 						for (Player partyPlayer : groupMembers)
@@ -659,6 +669,7 @@ public class Attackable extends Npc
 										}
 									}
 								}
+								
 								rewards.remove(partyPlayer); // Remove the Player from the Attackable rewards
 							}
 							else if (calculateDistance3D(partyPlayer) < Config.ALT_PARTY_RANGE)
@@ -793,7 +804,7 @@ public class Attackable extends Npc
 				getAI().notifyAction(Action.ATTACKED, attacker);
 				
 				// Calculate the amount of hate this attackable receives from this attack.
-				double hateValue = (damage * 100) / (getLevel() + 7);
+				long hateValue = ((long) damage * 100) / (getLevel() + 7);
 				if (skill == null)
 				{
 					hateValue *= attacker.getStat().getMul(Stat.HATE_ATTACK, 1);
@@ -1021,6 +1032,7 @@ public class Attackable extends Npc
 				break;
 			}
 		}
+		
 		if (found)
 		{
 			result.add(secondMostHated);
@@ -1047,6 +1059,7 @@ public class Attackable extends Npc
 			
 			result.add(ai.getAttacker());
 		}
+		
 		return result;
 	}
 	
@@ -1089,6 +1102,7 @@ public class Attackable extends Npc
 			ai.stopHate();
 			return 0;
 		}
+		
 		return ai.getHate();
 	}
 	
@@ -1135,6 +1149,7 @@ public class Attackable extends Npc
 					for (ItemHolder drop : deathItems)
 					{
 						final ItemTemplate item = ItemData.getInstance().getTemplate(drop.getId());
+						
 						// Check if the autoLoot mode is active
 						if (Config.AUTO_LOOT_ITEM_IDS.contains(item.getId()) || isFlying() || (!item.hasExImmediateEffect() && ((!_isRaid && Config.AUTO_LOOT) || (_isRaid && Config.AUTO_LOOT_RAIDS))))
 						{
@@ -1146,6 +1161,7 @@ public class Attackable extends Npc
 							{
 								SkillCaster.triggerCast(mainDamageDealer, null, skillHolder.getSkill(), null, false);
 							}
+							
 							mainDamageDealer.broadcastInfo(); // ? check if this is necessary
 						}
 						else
@@ -1157,6 +1173,7 @@ public class Attackable extends Npc
 							}
 						}
 					}
+					
 					deathItems.clear();
 				}
 			}
@@ -1175,6 +1192,7 @@ public class Attackable extends Npc
 			for (ItemHolder drop : deathItems)
 			{
 				final ItemTemplate item = ItemData.getInstance().getTemplate(drop.getId());
+				
 				// Check if the autoLoot mode is active
 				if (Config.AUTO_LOOT_ITEM_IDS.contains(item.getId()) || isFlying() || (!item.hasExImmediateEffect() && ((!_isRaid && Config.AUTO_LOOT) || (_isRaid && Config.AUTO_LOOT_RAIDS))) || (item.hasExImmediateEffect() && Config.AUTO_LOOT_HERBS))
 				{
@@ -1195,6 +1213,7 @@ public class Attackable extends Npc
 					broadcastPacket(sm);
 				}
 			}
+			
 			deathItems.clear();
 		}
 	}
@@ -1253,6 +1272,7 @@ public class Attackable extends Npc
 				lootItems.add(ItemData.getInstance().getTemplate(item.getId()));
 			}
 		}
+		
 		return lootItems;
 	}
 	
@@ -1287,8 +1307,10 @@ public class Attackable extends Npc
 			{
 				attacker.sendPacket(SystemMessageId.THE_CORPSE_IS_TOO_OLD_THE_SKILL_CANNOT_BE_USED);
 			}
+			
 			return true;
 		}
+		
 		return false;
 	}
 	
@@ -1305,8 +1327,10 @@ public class Attackable extends Npc
 			{
 				sweeper.sendPacket(SystemMessageId.THERE_ARE_NO_PRIORITY_RIGHTS_ON_A_SWEEPER);
 			}
+			
 			return false;
 		}
+		
 		return true;
 	}
 	
@@ -1338,6 +1362,7 @@ public class Attackable extends Npc
 			_overhitAttacker = null;
 			return;
 		}
+		
 		overhitEnabled(true);
 		_overhitDamage = overhitDmg;
 		_overhitAttacker = attacker;
@@ -1436,10 +1461,12 @@ public class Attackable extends Npc
 						break;
 					}
 				}
+				
 				xp *= mul;
 				sp *= mul;
 			}
 		}
+		
 		return new double[]
 		{
 			xp,
@@ -1533,6 +1560,7 @@ public class Attackable extends Npc
 			{
 				_champion = true;
 			}
+			
 			if (Config.SHOW_CHAMPION_AURA)
 			{
 				setTeam(_champion ? Team.RED : Team.NONE, false);
@@ -1644,6 +1672,7 @@ public class Attackable extends Npc
 			{
 				count += diff;
 			}
+			
 			_harvestItem.set(new ItemHolder(_seed.getCropId(), count * Config.RATE_DROP_MANOR));
 		}
 	}
@@ -1843,20 +1872,25 @@ public class Attackable extends Npc
 			{
 				_aggroList.remove(target);
 			}
+			
 			if (_aggroList.isEmpty())
 			{
 				if (getAI() instanceof AttackableAI)
 				{
 					((AttackableAI) getAI()).setGlobalAggro(-25);
 				}
+				
 				if (!isFakePlayer())
 				{
 					setWalking();
 				}
+				
 				clearAggroList();
 			}
+			
 			getAI().setIntention(Intention.ACTIVE);
 		}
+		
 		super.setTarget(object);
 	}
 }
